+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/rules/JustWithRule.kt
@@ -7,22 +7,25 @@ public class JustWithRule(
     private val justWith: List<String>
 ) : PopcornGuineaPigRule {
     override fun check(deps: List<InternalDependenciesMetadata>): ArchitectureViolationError? {
-        val internalProjectDependenciesSortedByModuleName = deps.sortedBy { it.moduleName }
-        val internalProjectDependencyModuleNames = internalProjectDependenciesSortedByModuleName.map { it.moduleName }
+        val affectedRelationships = mutableListOf<InternalDependenciesMetadata>()
 
-        val justWithSorted = justWith.sorted()
-        if (internalProjectDependencyModuleNames != justWithSorted) {
-            val affectedRelationships = internalProjectDependenciesSortedByModuleName.filterNot {
-                justWithSorted.contains(it.moduleName)
+        deps.forEach { dep ->
+            val matchesAnyPattern = justWith.any { pattern ->
+                pattern.toRegex().matches(dep.moduleName)
             }
+            if (matchesAnyPattern.not()) {
+                affectedRelationships.add(dep)
+            }
+        }
 
-            return ArchitectureViolationError(
+        return if (affectedRelationships.isNotEmpty()) {
+            ArchitectureViolationError(
                 message = "This module should depends on $justWith",
                 rule = this,
                 affectedRelationship = affectedRelationships
             )
+        } else {
+            null
         }
-
-        return null
     }
 }